Desenvolvendo funcionalidade de aprovação de propriedade #2

+++ app/Http/Controllers/PropertyImageController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
+use App\Models\PreapprovedPropertyImage;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use App\Models\PropertyImage;
@@ -12,17 +13,27 @@ class PropertyImageController extends Controller
     {
         $id = $request->input('id');
 
-        $image = PropertyImage::find($id);
+        if($request->type == 'property')
+            $image = PropertyImage::find($id);
+        elseif ($request->type == 'preapproved_property')
+            $image = PreapprovedPropertyImage::find($id);
+        else
+            return response()->json(['error' => 'Imagem não encontrada'], 404);
+
 
         if (!$image) {
             return response()->json(['error' => 'Imagem não encontrada'], 404);
         }
 
         // Remove do disco
-        if (Storage::disk('public')->exists($image->path)) {
+        if (Storage::disk('public')->exists($image->path) && $request->type != 'preapproved_property') {
             Storage::disk('public')->delete($image->path);
         }
 
+        $preapprovedImage = PreapprovedPropertyImage::where('path', $image->path)->first();
+        if($preapprovedImage){
+            $preapprovedImage->delete();
+        }
         // Remove do banco
         $image->delete();
